Update enumerated types SQL
