+++ apps/studio/data/enumerated-types/enumerated-type-create-mutation.ts
@@ -1,3 +1,4 @@
+import { getCreateEnumeratedTypeSQL } from '@supabase/pg-meta'
 import { wrapWithTransaction } from '@supabase/pg-meta/src/query'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { executeSql } from 'data/sql/execute-sql-query'
@@ -23,12 +24,7 @@ export async function createEnumeratedType({
   description,
   values,
 }: EnumeratedTypeCreateVariables) {
-  const createSql = `create type "${schema}"."${name}" as enum (${values
-    .map((x) => `'${x}'`)
-    .join(', ')});`
-  const commentSql =
-    description !== undefined ? `comment on type "${schema}"."${name}" is '${description}';` : ''
-  const sql = wrapWithTransaction(`${createSql} ${commentSql}`)
+  const sql = getCreateEnumeratedTypeSQL({ schema, name, values, description })
   const { result } = await executeSql({ projectRef, connectionString, sql })
   return result
 }
