+++ packages/pg-meta/src/sql/studio/database/index.ts
@@ -13,5 +13,6 @@ export * from './queues'
 export * from './rows'
 export * from './table-definition'
 export * from './triggers'
+export * from './types'
 export * from './vault'
 export * from './views'
